Fix slider disable opacity
