+++ forui/lib/src/foundation/portal/portal_shift.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
-
-import 'package:forui/src/foundation/alignment.dart';
+import 'package:forui/src/foundation/rendering.dart';
 
 /// A portal's target.
 typedef FPortalTarget = ({Offset offset, Size size, Alignment anchor});
