+++ forui/test/src/foundation/rendering_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:forui/src/foundation/alignment.dart';
+import 'package:forui/src/foundation/rendering.dart';
 
 void main() {
   group('Alignments', () {
