+++ forui/test/src/theme/color_scheme_test.dart
@@ -9,6 +9,7 @@ void main() {
   group('FColorScheme', () {
     const scheme = FColorScheme(
       brightness: Brightness.light,
+      disabledColorLightness: 0.5,
       background: Colors.black,
       foreground: Colors.black12,
       primary: Colors.black26,
@@ -30,6 +31,7 @@ void main() {
       test('all arguments', () {
         final copy = scheme.copyWith(
           brightness: Brightness.dark,
+          disabledColorLightness: 0.75,
           background: Colors.red,
           foreground: Colors.greenAccent,
           primary: Colors.yellow,
@@ -46,6 +48,7 @@ void main() {
         );
 
         expect(copy.brightness, equals(Brightness.dark));
+        expect(copy.disabledColorLightness, equals(0.75));
         expect(copy.background, equals(Colors.red));
         expect(copy.foreground, equals(Colors.greenAccent));
         expect(copy.primary, equals(Colors.yellow));
@@ -70,6 +73,7 @@ void main() {
         builder.properties.map((p) => p.toString()),
         [
           EnumProperty('brightness', Brightness.light),
+          DoubleProperty('disabledColorLightness', 0.5),
           ColorProperty('background', Colors.black),
           ColorProperty('foreground', Colors.black12),
           ColorProperty('primary', Colors.black26),
